requirements.txt
yfinance
pandas
sqlalchemy
psycopg2-binary
dbt-postgres
python-dotenv